+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added:
  - User sessions history
  - Profile setting `excludeFromSearch($bool)` to exclude your account from search results
+ - Profile setting `setLocal($locale)`
 
 ## v5.3.11 - 2017-05-23
 ### Fixed: